// src/app/models/customer.ts
export interface Country {
    name?: string;
    code?: string;
}

export interface Representative {
    name?: string;
    image?: string;
}

export interface Customer {
    id?: number;
    name?: string;
    country?: Country;
    company?: string;
    date: Date;
    status?: string;
    activity?: number;
    representative?: Representative;
}